Add `ed25519` signature provider (#17)
* Move `SignatureProvider` to the abstractions assembly.

This is so it can be used by a separate library for BouncyCastle (needed for ed25519) without requiring a dependency on the NSign.SignatureProviders assembly.

* Add the NSign.BouncyCastle assembly

Also add a signature provider for EdDSA using curve ed25519, and corresponding unit tests.

* Fix typos in signature providers.

* Update package dependencies

* Update README to include reference to NSign.BouncyCastle package.

* Bump revision

* Fix the unit tests for the ed25519 signature provider

Use PEM files for private and public keys so the tests work on Linux too. Also update the package README to have minimal instructions on how to extract private and public keys from .pfx and .cer files.